+++ FluidNC/src/Machine/Macros.cpp
@@ -70,8 +70,8 @@ bool Macro::run(Channel* channel) {
     return false;
 }
 
-Error MacroChannel::readLine(char* line, int maxlen) {
-    int                len       = 0;
+Error MacroChannel::readLine(char* line, size_t maxlen) {
+    size_t             len       = 0;
     const std::string& gcode     = _macro->_gcode;
     const int          gcode_len = gcode.length();
     while (_position < gcode_len) {
