+++ FluidNC/src/MotionControl.h
@@ -33,11 +33,11 @@ void mc_arc(float*            target,
             float*            position,
             float*            offset,
             float             radius,
-            size_t            axis_0,
-            size_t            axis_1,
-            size_t            axis_linear,
+            uint8_t           axis_0,
+            uint8_t           axis_1,
+            uint8_t           axis_linear,
             bool              is_clockwise_arc,
-            int               pword_rotations);
+            uint32_t          pword_rotations);
 
 // Dwell for a specific number of seconds
 bool mc_dwell(int32_t milliseconds);
